Moved millis() and delay() to mock_ticks.h
