+++ firmware/boron-ins-fsm/test/base.h
@@ -10,6 +10,8 @@
 #include "mocks/mock_System.h"
 #include "mocks/mock_os_queue_t.h"
 #include "mocks/mock_ble.h"
+#include "mocks/mock_logging.h"
+#include "mocks/mock_ticks.h"
 
 // Particle.h library include files
 // Copied from local Particle toolchain files
